:arrow_right: Move all test (*.spec.js) files to __test__ subfolders
