+++ src/app/components/basic/__tests__/Wrapper.spec.js
@@ -1,5 +1,5 @@
 import { Component } from '@local/system'
-import Wrapper from './Wrapper';
+import Wrapper from '../Wrapper';
 
 describe('constructor', () => {
     it('stores all components in state', () => {
